Fix the build for Windows, mostly due to having to link against a few Boost libraries.
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@1015 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/formats/powertab_old/powertabdocument/score.cpp
@@ -730,7 +730,7 @@ Score::ChordDiagramPtr Score::GetChordDiagram(uint32_t index) const
 /// Determines if a guitar index is valid
 /// @param index guitar index to validate
 /// @return True if the guitar index is valid, false if not
-bool Score::IsValidGuitarIndex(uint32_t index) const
+bool Score::IsValidGuitarIndex(size_t index) const
 {
     return index < GetGuitarCount();
 }
@@ -745,7 +745,7 @@ size_t Score::GetGuitarCount() const
 /// Gets the nth guitar in the score
 /// @param index Index of the guitar to get
 /// @return The nth guitar in the score
-Score::GuitarPtr Score::GetGuitar(uint32_t index) const
+Score::GuitarPtr Score::GetGuitar(size_t index) const
 {
     PTB_CHECK_THAT(IsValidGuitarIndex(index), GuitarPtr());
     return m_guitarArray[index];
